pyproject.toml
[project]
name = "sre-agent"
version = "0.2.0"
description = "SRE Agent - Google Cloud Observability analysis agent for traces, logs, and metrics"
authors = [{ name = "SRT", email = "srtuxx@gmail.com" }]
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
    "google-cloud-trace>=1.0.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "google-auth>=2.18.1",
    "grpcio>=1.63.0",

    # Phase 4: Extended GCP integrations
    "google-cloud-logging>=3.10.0",
    "google-cloud-monitoring>=2.21.0",
    "google-cloud-error-reporting>=1.14.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",

    # Log pattern analysis
    "drain3>=0.9.11",
]

requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
    "pytest-xdist>=3.8.0",
    "flake8>=7.1.1",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501", "C901"]

[tool.ruff.lint.isort]
known-first-party = ["sre_agent"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""



[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.agent-starter-pack]
example_question = "Analyze traces and logs in project my-gcp-project to find performance issues"

[tool.agent-starter-pack.settings]
agent_directory = "sre_agent"
